day5_enterprise/requirements.txt
fastapi
email-validator>=2.0
uvicorn
streamlit
requests
langchain
langchain-openai
rank_bm25
numpy
tiktoken
python-dotenv